API

API documentation

QuaternionIntegrator.rotate — Function.

rotate(q::Quaternion, v::Vector)

Apply orientation quaternion q to rotate vector v, return rotated vector.

An object's orientation quaternion rotates vectors from the body-fixed coordinates to world coordinates. For the opposite rotation, call rotate(inv(q), v).

source

QuaternionIntegrator.orientation — Function.

orientation(v::Vector, angle::Real)

Return an orientation quaternion with the given axis v and rotation angle.

The vector v is assumed to be normalized to unit length.

source

QuaternionIntegrator.integrate — Function.

integrate(q0::Quaternion, ω0::Vector, Ib::Matrix, ∆t, torque::Function)

Integrate a rotational state ahead by one time step.

q0, current orientation quaternion.

ω0, current angular velocity (length-3 vector).

Ib, inertial tensor of the object in body coordinates (3×3 matrix).

∆t, length of time step.

torque, a function that returns a (length-3) torque vector given an orientation.

Returns (q1, ω1), the orientation and angular velocity at the end of the time step.

Using the Unitful package, all inputs can have units, all the mathematics are done with units and the return values will have correct units.

q0 is dimensionless.

ω0 has dimension of 1/time (SI: 1/second)

Ib has dimension mass * length^2 (SI: kilogram * meter^2)

∆t has dimension of time (SI: second)

torque returns a value with dimension of torque (SI: Newton * meter)

source

integrate(q0::Quaternion, ω0::Vector, Ib::Matrix, ∆t, torque::Function, N::Integer)

Integrate a rotational state ahead by N time steps.

Returns (q1, ω1), the orientation and angular velocity after N integration steps.

source